Added showDetails function and event listener to Pokemon buttons
Added the showDetails function to log all Pokemon details to the console when a Pokemon button is clicked. Implemented an event listener in the addListItem function to call showDetails with the corresponding Pokemon object as a parameter upon button click. This allows for further actions or functionalities to be performed based on the Pokemon details.