Added Promise Polyfills and Fetch Functionality
Added Promise polyfills to ensure compatibility with older browsers and implemented functionality to fetch data from an external source for the project. Removed the array of Pokémon objects in the "scripts.js" file and replaced it with an empty array.

Implemented functions in the data repository to load data from an external source. Added a LoadList() function that uses fetch to GET the complete list of Pokémon from the PokeAPI. Used the add() function to add each Pokémon from the results to the pokemonList variable, setting name and detailsUrl as keys. Also added a loadDetails() function to GET the Pokémon details using the URL from the Pokémon object as a parameter.

Upon completion of the GET request, utilized then to return a JSON response and assigned some of the details received from the response to the Pokémon in the pokemonList, including imgUrl and height.

Ensured that both LoadList() and loadDetails() functions are assigned to keys with the same name in the returned object of the pokemonRepository. Also ensured that the Pokémon list is rendered only after receiving all information from the server.

Called the LoadList() function and executed getAll from the pokemonRepository. For each Pokémon, called the addListItem() function from the previous Exercise.

Updated the showDetails() function to load Pokémon details from the API instead of static data. Within the showDetails() function, called the loadDetails() function and passed the Pokémon object as a parameter. Logged the result in the console.

This commit enhances the project by incorporating Promise polyfills for broader browser compatibility and integrating fetching functionality to retrieve Pokémon data from an external API.